+++ .gitignore
@@ -45,3 +45,4 @@ qtcreator-*
 
 # Catkin custom files
 CATKIN_IGNORE
+cfg/darkflow_config/ckpt/*
