+++ launch/darkflow_ros.launch
@@ -10,15 +10,15 @@
   <arg name="yolo_checkpoint"   default="-1"/>
  
   <!-- Config Cameras -->
-  <arg name="camera_0" default="/camera/rgb/image_raw"/>
-  <arg name="camera_topic_0" default="/camera/rgb/camera_info"/>
+  <arg name="camera" default="/camera/rgb/image_raw"/>
+  <arg name="camera_topic" default="/camera/rgb/camera_info"/>
 
   <!-- Load parameters -->
   <rosparam command="load" ns="darkflow_ros" file="$(find iarc7_vision_cnn)/cfg/ros.yaml"/>
   <rosparam command="load" ns="darkflow_ros" file="$(find iarc7_vision_cnn)/cfg/yolov2-tiny.yaml"/>
 
   <!-- Start darkflow and ros wrapper -->
-  <node pkg="iarc7_vision_cnn" type="darkflow_ros.py" name="darkflow_ros" output="screen" launch-prefix="$(arg launch_prefix)">
+  <node pkg="iarc7_vision_cnn" type="DarkflowRos.py" name="darkflow_ros" output="screen" launch-prefix="$(arg launch_prefix)">
 
     <param name="checkpoint_path" value="$(arg yolo_checkpoint_path)" />
     <param name="config_path"     value="$(arg yolo_config_path)" />
@@ -27,8 +27,8 @@
     <param name="gpu_usage"     value="$(arg yolo_gpu_usage)" />
     <param name="checkpoint"     value="$(arg yolo_checkpoint)" />
 
-    <remap from="image0" to="$(arg camera_0)"/>
-    <remap from="itopic0" to="$(arg camera_topic_0)"/>
+    <remap from="image" to="$(arg camera)"/>
+    <remap from="camera_topic" to="$(arg camera_topic)"/>
     
   </node>
